+++ src/BifReturnVal.cc
@@ -1,0 +1,11 @@
+// See the file "COPYING" in the main distribution directory for copyright.
+
+#include "BifReturnVal.h"
+#include "Val.h"
+
+BifReturnVal::BifReturnVal(std::nullptr_t) noexcept
+	{}
+
+BifReturnVal::BifReturnVal(Val* v) noexcept
+	: rval(AdoptRef{}, v)
+	{}
